pyproject.toml
[project]
name = "missouri-vsr"
version = "0.1.0"
description = ""
authors = [{ name = "David Eads", email = "deads@themarshallproject.org" }]
requires-python = ">=3.12,<3.13"
readme = "README.md"
dependencies = [
    "dagster>=1.10.5,<2",
    "dagster-webserver>=1.10.5,<2",
    "pandas>=2.2.3,<3",
    "pyairtable>=3.0.2,<4",
    "boto3>=1.37.12,<2",
    "markdown~=3.7",
    "google-api-python-client>=2.166.0,<3",
    "python-slugify>=8.0.4",
    "camelot-py>=1.0.0",
    "pypdf2>=3.0.1",
    "openpyxl>=3.1.5",
    "pyarrow>=16.0.0",
    "rapidfuzz>=3.14.3",
    "geocodio-library-python>=0.3.0",
    "matplotlib>=3.10.8",
    "httpx>=0.28.1",
    "geopandas>=1.1.2",
    "shapely>=2.1.2",
    "pyogrio>=0.12.1",
    "duckdb>=1.2.0",
]

[project.scripts]
vsr-parse = "missouri_vsr.cli.cli:main"

[tool.dagster]
module_name = "missouri_vsr.definitions"
code_location_name = "missouri_vsr"

[tool.hatch.build.targets.sdist]
include = ["missouri_vsr"]

[tool.hatch.build.targets.wheel]
include = ["missouri_vsr"]

[tool.hatch.build.targets.wheel.sources]
"missouri_vsr" = "missouri_vsr"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "altair>=6.0.0",
    "ipdb>=0.13.13",
    "marimo>=0.19.7",
    "pytest>=9.0.2",
]

[tool.pytest.ini_options]
markers = [
    "network: tests that make live network calls",
]
